+++ package.json
@@ -19,6 +19,7 @@
     "body-parser": "^1.18.2",
     "corser": "^2.0.1",
     "express": "^4.16.2",
+    "jsonwebtoken": "^8.1.0",
     "morgan": "^1.9.0",
     "mysql": "^2.15.0",
     "sha256": "^0.2.0"
