+++ swagger/api/swagger/swagger.yaml
@@ -43,6 +43,10 @@ paths:
           in: query
           required: true
           type: integer
+        - name: token
+          in: query
+          required: false
+          type: string
       responses:
         "200":
           description: Success
